Restructure project as a proper Python package
- Create llm_eti package directory with __init__.py
- Move all core Python modules into the package
- Update all imports throughout the codebase to use package structure
- Update pyproject.toml to properly define the package
- Add package import tests
- Enable clean imports for testing (from llm_eti import ...)

This restructuring makes the project more maintainable and allows
for cleaner test imports without path manipulation.
