+++ book/scripts/run_gruber_saez_simulation.py
@@ -11,9 +11,9 @@ from pathlib import Path
 # Add parent directory to path to import from main project
 sys.path.append(str(Path(__file__).parent.parent.parent))
 
-from simulation_engine import TaxSimulation
-from gpt_utils import GPTClient
-from config import SimulationParams
+from llm_eti.simulation_engine import TaxSimulation
+from llm_eti.gpt_utils import GPTClient
+from llm_eti.config import SimulationParams
 
 
 def main():
