+++ pyproject.toml
@@ -46,6 +46,9 @@ pdf = [
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
+[tool.hatch.build.targets.wheel]
+packages = ["llm_eti"]
+
 [tool.ruff]
 line-length = 88
 select = ["E", "F", "I"]
