+++ requirements.txt
@@ -1,0 +1,17 @@
+openai>=1.0.0
+pandas>=2.0.0
+numpy>=1.24.0
+matplotlib>=3.7.0
+seaborn>=0.12.0
+python-dotenv>=1.0.0
+click>=8.1.0
+tqdm>=4.65.0
+scikit-learn>=1.3.0
+statsmodels>=0.14.0
+jupyter-book>=0.15.0
+sphinx>=4.0
+pydata-sphinx-theme>=0.13.0
+myst-nb>=0.17.0
+sphinx-book-theme>=1.0.0
+sphinx-design>=0.4.0
+streamlit>=1.28.0
