+++ tests/test_simulation.py
@@ -1,6 +1,6 @@
 import pytest
-from simulation_engine import SimulationParams, TaxSimulation
-from gpt_utils import GPTClient
+from llm_eti.simulation_engine import SimulationParams, TaxSimulation
+from llm_eti.gpt_utils import GPTClient
 
 
 class MockGPTClient:
